refactor: Replace Helper with Trait.
